+++ R/ip_checks.R
@@ -1,0 +1,42 @@
+#' IP checks
+#'
+#' @description
+#' Quick ways to get your IP address from within R
+#'
+#' @examples
+#' # determine IPV4
+#' get_ip()
+#'
+#'
+#' @name ip
+NULL
+
+#' @rdname ip
+#' @export
+get_ip <- function(){
+
+  ip <- NA
+
+  if(is.os_win()){
+
+    systext <- system("ipconfig", intern=TRUE)
+
+    ip <- gsub(".*? ([[:digit:]])", "\\1", systext[grep("IPv4", systext)])
+
+
+
+  } else if(is.os_lnx() || is.os_unx()) {
+
+    systext <- system("ifconfig", intern=TRUE)
+
+    ip <- gsub(".*? ([[:digit:]])", "\\1", systext[grep("inet", systext)])
+
+  } else if(is.os_mac()){
+
+  } else{
+    stop("Couldn't detect your operating system or IP address from system calls")
+  }
+
+  return(ip)
+
+}