+++ NEWS.md
@@ -4,3 +4,5 @@
 * Added a `NEWS.md` file to track changes to the package.
 * New `get_os()`, `is.os_mac`, `is.os_win`, `is.os_lnx`, and `is.os_unx` helper
   functions to quickly determine operating system.
+* New `n_unique()` function to count unique items.
+* New `get_most_frequent()` function to get most frequent thing(s) in x